Input driven upload type decision
